# doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Spokane Economic Dashboard",
        "description": "API Endpoints for interactions between the MongoDB cluster and the Census Bureau API",
        "version": "1.0.0",
        "contact": {
            "email": "zfoteff@protonmail.com",
            "name": "CS07 Team"
        }
    },
...
    "host": "spokaneeconomicdashboard.herokuapp.com",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/button": {
            "get": {
                "tags": ["nav"],
                "summary": "Render 'The Button' page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the button page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            }
        },
        "/endpoints": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the manual input indicator update page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the button page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            }
        },
        "/auth": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the admin page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the admin page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the login page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the login page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Auth"],
                "summary": "Submit a username and password to be verified by the application",
                "description": "Endpoint interacts with MongoDb to retrieve a user and it's stored password and match it to the inputted values. If there is a match, redirect the user to the ",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "username": {
                                "example": "USERNAME@ALLOWED_DOMAIN.com"
                            },
                            "password": {
                                "example": "password"
                            }
                        }
                    }
                }],
                "responses": {
                    "302": {
                        "description": "Successful login. Redirect to admin page"
                    },
                    "403": {
                        "description": "Bad login. Reload login page and try again"
                    }
                }
            }
        },
        "/register": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the register page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the register page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Auth"],
                "summary": "Register new authorized email to access the admin page",
                "description": "Endpoint checks if the user's email is valid, then it saves the user's username and redirects them to the admin page",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "username": {
                                "example": "USERNAME@ALLOWED_DOMAIN.com"
                            },
                            "password": {
                                "example": "password"
                            }
                        }
                    }
                }],
                "responses": {
                    "302": {
                        "description": "Successfully registered user. Redirect to admin page"
                    },
                    "400": {
                        "description": "Failed to register user"
                    }
                }
            }
        },
        "/change": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the change password page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the login page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Auth"],
                "summary": "Submit the user's email to the change password service",
                "description": "Checks if the user is a registered user of the application, then it emails the user a pin code and redirects them to the pin verification page",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "example": "any"
                            }
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/pin": {
            "get": {
                "tags": ["nav"],
                "summary": "Render the verify pin page for the user",
                "responses": {
                    "200": {
                        "description": "Success. Render the verify pin page",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "examples": {
                            "text/html": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Auth"],
                "summary": "Submit a pin and new password.",
                "description": "The endpoint verifies that the submitted pin matches the pin sent to the user and that it was submitted in the proper amount of time. Then it will update the requesting user's password with the new value",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "pin": {
                                "example": "123456"
                            },
                            "updatedpass": {
                                "example": "new password"
                            }
                        }
                    }
                }],
                "responses": {
                    "302": {
                        "description": "Successfully verified pin and changed password. Redirect to admin page"
                    },
                    "401": {
                        "description": "Failed to verify pin. Redirect to change password page"
                    }
                }
            }
        },
...
    }
}